Added keyboard control to lobby. Added Fps namespace to scripts.
