+++ workers/unity/Assets/Fps/Prefabs/UI/Tables/LobbyTableEntry.prefab
@@ -78,7 +78,7 @@ MonoBehaviour:
     m_HorizontalOverflow: 0
     m_VerticalOverflow: 0
     m_LineSpacing: 1
-  m_Text: DEPLOYMENT
+  m_Text: TEMPLATE DO NOT REMOVE
 --- !u!1 &3617305780785071335
 GameObject:
   m_ObjectHideFlags: 0
@@ -341,7 +341,6 @@ MonoBehaviour:
   Columns:
   - {fileID: 223568142224586518}
   - {fileID: 3107015718779868864}
-  Id: 0
 --- !u!114 &4799642357629010387
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -355,7 +354,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   m_Navigation:
-    m_Mode: 3
+    m_Mode: 0
     m_SelectOnUp: {fileID: 0}
     m_SelectOnDown: {fileID: 0}
     m_SelectOnLeft: {fileID: 0}
