+++ workers/unity/Assets/Fps/Scripts/UI/Controllers/MatchmakingScreenController.cs
@@ -1,61 +1,62 @@
-﻿using Fps;
-using UnityEngine;
-using UnityEngine.UI;
+﻿using UnityEngine;
 
-public class MatchmakingScreenController : MonoBehaviour
+namespace Fps
 {
-    public InputFieldPlayerName PlayerNameInputField;
-    public FpsUIButton QuickJoinButton;
-    public FpsUIButton BrowseButton;
-    public FrontEndUIController frontEndUIController;
-
-    private const int quickJoinDefaultIndex = 0;
-    private const int quickJoinCancelIndex = 1;
-    private const int quickJoinJoiningIndex = 2;
-
-    private void Awake()
+    public class MatchmakingScreenController : MonoBehaviour
     {
-        frontEndUIController = GetComponentInParent<FrontEndUIController>();
-        PlayerNameInputField.OnNameChanged += SetButtonsEnabled;
-        Debug.Assert(PlayerNameInputField != null);
-        Debug.Assert(BrowseButton != null);
-        Debug.Assert(QuickJoinButton != null);
-    }
+        public InputFieldPlayerName PlayerNameInputField;
+        public FpsUIButton QuickJoinButton;
+        public FpsUIButton BrowseButton;
+        public FrontEndUIController frontEndUIController;
 
-    private void SetButtonsEnabled(bool isValid)
-    {
-        isValid = true;
-        Debug.LogFormat("<b>Setting button enableds to " + isValid + "</b>");
-        QuickJoinButton.enabled = isValid;
-        BrowseButton.enabled = isValid;
-    }
+        private const int quickJoinDefaultIndex = 0;
+        private const int quickJoinCancelIndex = 1;
+        private const int quickJoinJoiningIndex = 2;
 
-    //TODO Move these to anim?
-    public void FindingQuickGame()
-    {
-        BrowseButton.enabled = false;
-        QuickJoinButton.SetText(quickJoinCancelIndex);
-    }
+        private void Awake()
+        {
+            frontEndUIController = GetComponentInParent<FrontEndUIController>();
+            PlayerNameInputField.OnNameChanged += SetButtonsEnabled;
+            Debug.Assert(PlayerNameInputField != null);
+            Debug.Assert(BrowseButton != null);
+            Debug.Assert(QuickJoinButton != null);
+        }
 
-    public void FoundQuickGame()
-    {
-        QuickJoinButton.enabled = false;
-        QuickJoinButton.SetText(quickJoinJoiningIndex);
-    }
+        private void SetButtonsEnabled(bool isValid)
+        {
+            isValid = true;
+            Debug.LogFormat("<b>Setting button enableds to " + isValid + "</b>");
+            QuickJoinButton.enabled = isValid;
+            BrowseButton.enabled = isValid;
+        }
 
-    public void CancelQuickJoin()
-    {
-        QuickJoinButton.SetText(quickJoinDefaultIndex);
-        SetButtonsEnabled(true);
-    }
+        //TODO Move these to anim?
+        public void FindingQuickGame()
+        {
+            BrowseButton.enabled = false;
+            QuickJoinButton.SetText(quickJoinCancelIndex);
+        }
 
-    public void QuickJoinPressed()
-    {
+        public void FoundQuickGame()
+        {
+            QuickJoinButton.enabled = false;
+            QuickJoinButton.SetText(quickJoinJoiningIndex);
+        }
 
-    }
+        public void CancelQuickJoin()
+        {
+            QuickJoinButton.SetText(quickJoinDefaultIndex);
+            SetButtonsEnabled(true);
+        }
 
-    public void BrowsePressed()
-    {
-        frontEndUIController.SetScreenTo(FrontEndUIController.ScreenType.Lobby);
+        public void QuickJoinPressed()
+        {
+
+        }
+
+        public void BrowsePressed()
+        {
+            frontEndUIController.SetScreenTo(FrontEndUIController.ScreenType.Lobby);
+        }
     }
 }
