+++ workers/unity/Assets/Fps/Scripts/UI/Controllers/ResultsScreenController.cs
@@ -1,91 +1,94 @@
 ﻿using UnityEngine;
 
-public class ResultsScreenController : MonoBehaviour
+namespace Fps
 {
-    [SerializeField] private GameObject rankDivider;
-
-    public Color BackgroundStripColor1;
-    public Color BackgroundStripColor2;
-    public Color LocalPlayerBackgroundColor;
-    public Color LocalPlayerTextColor;
-    public Color DefaultTextColor;
-
-    private const int MaxVisibleRowCount = 12; // Includes header
-    private Table resultsTable;
-    private Color currentBgColor;
-
-    private void Awake()
+    public class ResultsScreenController : MonoBehaviour
     {
-        resultsTable = GetComponentInChildren<Table>();
-    }
+        [SerializeField] private GameObject rankDivider;
 
-    public void SetResults(ResultsData[] results)
-    {
-        resultsTable.ClearEntries();
-        currentBgColor = BackgroundStripColor1;
+        public Color BackgroundStripColor1;
+        public Color BackgroundStripColor2;
+        public Color LocalPlayerBackgroundColor;
+        public Color LocalPlayerTextColor;
+        public Color DefaultTextColor;
 
-        var playerRank = Random.Range(0, results.Length); // TODO Replace with players rank from game
+        private const int MaxVisibleRowCount = 12; // Includes header
+        private Table resultsTable;
+        private Color currentBgColor;
 
-        AdjustTableHeight(playerRank);
-
-        for (var i = 0; i < Mathf.Min(10, results.Length); i++)
+        private void Awake()
         {
-            AddPlayerToTable(results[i], playerRank == i);
-            currentBgColor = i % 2 == 0 ? BackgroundStripColor2 : BackgroundStripColor1;
+            resultsTable = GetComponentInChildren<Table>();
         }
 
-        if (playerRank >= 10)
+        public void SetResults(ResultsData[] results)
         {
-            AddPlayerToTable(results[playerRank], true);
+            resultsTable.ClearEntries();
+            currentBgColor = BackgroundStripColor1;
+
+            var playerRank = Random.Range(0, results.Length); // TODO Replace with players rank from game
+
+            AdjustTableHeight(playerRank);
+
+            for (var i = 0; i < Mathf.Min(10, results.Length); i++)
+            {
+                AddPlayerToTable(results[i], playerRank == i);
+                currentBgColor = i % 2 == 0 ? BackgroundStripColor2 : BackgroundStripColor1;
+            }
+
+            if (playerRank >= 10)
+            {
+                AddPlayerToTable(results[playerRank], true);
+            }
         }
-    }
 
-    private void AdjustTableHeight(int playerRank)
-    {
-        if (playerRank < 10)
+        private void AdjustTableHeight(int playerRank)
         {
-            var rect = resultsTable.GetComponent<RectTransform>();
-            rect.offsetMin = new Vector2(rect.offsetMin.x, (MaxVisibleRowCount - 1) * resultsTable.EntryHeight * -.5f);
-            rect.offsetMax = new Vector2(rect.offsetMax.x, (MaxVisibleRowCount - 1) * resultsTable.EntryHeight * .5f);
-            rankDivider.SetActive(false);
+            if (playerRank < 10)
+            {
+                var rect = resultsTable.GetComponent<RectTransform>();
+                rect.offsetMin = new Vector2(rect.offsetMin.x, (MaxVisibleRowCount - 1) * resultsTable.EntryHeight * -.5f);
+                rect.offsetMax = new Vector2(rect.offsetMax.x, (MaxVisibleRowCount - 1) * resultsTable.EntryHeight * .5f);
+                rankDivider.SetActive(false);
+            }
+            else
+            {
+                var rect = resultsTable.GetComponent<RectTransform>();
+                rect.offsetMin = new Vector2(rect.offsetMin.x, MaxVisibleRowCount * resultsTable.EntryHeight * -.5f);
+                rect.offsetMax = new Vector2(rect.offsetMax.x, MaxVisibleRowCount * resultsTable.EntryHeight * .5f);
+                rankDivider.SetActive(true);
+            }
         }
-        else
+
+
+        private void AddPlayerToTable(ResultsData data, bool isLocalPlayer)
         {
-            var rect = resultsTable.GetComponent<RectTransform>();
-            rect.offsetMin = new Vector2(rect.offsetMin.x, MaxVisibleRowCount * resultsTable.EntryHeight * -.5f);
-            rect.offsetMax = new Vector2(rect.offsetMax.x, MaxVisibleRowCount * resultsTable.EntryHeight * .5f);
-            rankDivider.SetActive(true);
+            var entry = resultsTable.AddEntry();
+            entry.SetTextOnColumn(0, data.Rank.ToString());
+            entry.SetTextOnColumn(1, data.PlayerName);
+            entry.SetTextOnColumn(2, data.Kills.ToString());
+            entry.SetTextOnColumn(3, data.Deaths.ToString());
+            entry.SetTextOnColumn(4, data.KillDeathRatio.ToString("N2"));
+            entry.SetAllTextVisuals(isLocalPlayer ? LocalPlayerTextColor : DefaultTextColor, isLocalPlayer);
+            entry.SetBackgroundColor(isLocalPlayer ? LocalPlayerBackgroundColor : currentBgColor);
         }
-    }
 
-
-    private void AddPlayerToTable(ResultsData data, bool isLocalPlayer)
-    {
-        var entry = resultsTable.AddEntry();
-        entry.SetTextOnColumn(0, data.Rank.ToString());
-        entry.SetTextOnColumn(1, data.PlayerName);
-        entry.SetTextOnColumn(2, data.Kills.ToString());
-        entry.SetTextOnColumn(3, data.Deaths.ToString());
-        entry.SetTextOnColumn(4, data.KillDeathRatio.ToString("N2"));
-        entry.SetAllTextVisuals(isLocalPlayer ? LocalPlayerTextColor : DefaultTextColor, isLocalPlayer);
-        entry.SetBackgroundColor(isLocalPlayer ? LocalPlayerBackgroundColor : currentBgColor);
-    }
-
-    public class ResultsData
-    {
-        public readonly int Rank;
-        public readonly string PlayerName;
-        public readonly int Kills;
-        public readonly int Deaths;
-        public readonly float KillDeathRatio;
-
-        public ResultsData(int rank, string playerName, int kills, int deaths)
+        public class ResultsData
         {
-            Rank = rank;
-            PlayerName = playerName;
-            Kills = kills;
-            Deaths = deaths;
-            KillDeathRatio = deaths == 0 ? kills : kills / (float) deaths;
+            public readonly int Rank;
+            public readonly string PlayerName;
+            public readonly int Kills;
+            public readonly int Deaths;
+            public readonly float KillDeathRatio;
+
+            public ResultsData(int rank, string playerName, int kills, int deaths)
+            {
+                Rank = rank;
+                PlayerName = playerName;
+                Kills = kills;
+                Deaths = deaths;
+                KillDeathRatio = deaths == 0 ? kills : kills / (float) deaths;
+            }
         }
     }
 }
