+++ workers/unity/Assets/Fps/Scripts/UI/Controllers/Testers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 52d9cfdc5f0e8f04ba2a84ae748a2312
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
